requirements.txt
Flask==1.1.1
pandas==0.25.1
requests==2.22.0
Flask_Cors==3.0.8
Werkzeug==0.16.0
numpy==1.16.5
xlrd==1.2.0
Keras==2.3.1
dependency_injector==3.15.6
matplotlib==3.1.1
jsonpickle==1.3
PyYAML==5.3.1
scikit_learn==0.23.1
scipy==1.4.1
six==1.14.0
tensorflow==2.2.0
tensorflow-hub==0.9.0
six==1.14.0
Pillow==6.0.0